requirements.txt
python-dotenv==0.21.1
fastapi
openai
uvicorn
langchain
redis